RenderContext rerender
