+++ crates/zoon/src/component.rs
@@ -1,8 +1,10 @@
-use crate::element::{Element, RenderContext, IntoElement};
+use crate::element::{Element, IntoElement};
+use crate::render_context::RenderContext;
 use crate::tracked_call_stack::__TrackedCallStack;
 use crate::tracked_call::{__TrackedCall, TrackedCallId};
 use crate::render;
 use crate::runtime::CVARS;
+use crate::block_call_stack::{__Block, __BlockCallStack};
 use std::rc::Rc;
 use std::cell::RefCell;
 use crate::log;
@@ -17,7 +19,7 @@ pub struct Cmp {
 
 impl<'a> Element for Cmp {
     #[render]
-    fn render(&mut self, rcx: RenderContext) {
+    fn render(&mut self, mut rcx: RenderContext) {
         // log!("CMP render: {:#?}", TrackedCallId::current());
 
         // let aaa = __TrackedCallStack::last();
@@ -28,26 +30,29 @@ impl<'a> Element for Cmp {
 
         // log!("cmp render context: {:#?}", context);
 
-        if let Some(component_data_id) = self.component_data_id {
-            CVARS.with(move |c_vars| {
-                let mut c_vars = c_vars.borrow_mut();
+        let component_data_id = self.component_data_id.expect("component_data_id");
 
-                let mut component_data = c_vars.remove::<__ComponentData>(&component_data_id);
+        CVARS.with(move |c_vars| {
+            let mut c_vars = c_vars.borrow_mut();
 
-                component_data.rcx = Some(rcx);
-                component_data.parent_call = __TrackedCallStack::parent();
-                component_data.parent_selected_index = component_data.parent_call.as_ref().map(|parent| {
-                    parent.borrow().selected_index
-                });
+            let mut component_data = c_vars.remove::<__ComponentData>(&component_data_id);
 
-                // log!("PARENTTTTT CMP RENDER: {:#?}", component_data.parent_call);
-
-                // log!("CCCCCCCCCCCCCC component_data_id: {:#?}", component_data_id);
-                c_vars.insert(component_data_id, component_data);
-
+            component_data.rcx = Some(rcx);
+            component_data.parent_call = __TrackedCallStack::parent();
+            component_data.parent_selected_index = component_data.parent_call.as_ref().map(|parent| {
+                parent.borrow().selected_index
             });
-        }
-        self.element.render(rcx)
+
+            // log!("PARENTTTTT CMP RENDER: {:#?}", component_data.parent_call);
+
+            // log!("CCCCCCCCCCCCCC component_data_id: {:#?}", component_data_id);
+            c_vars.insert(component_data_id, component_data);
+        });
+
+        
+        rcx.component_id = Some(component_data_id);
+        self.element.render(rcx);
+
         // match self {
         //     Cmp::Element(element) => {
         //         element.render(rcx)
