+++ crates/zoon/src/render_context.rs
@@ -1,0 +1,28 @@
+use crate::{TrackedCallId, l_var::LVar};
+use crate::Node;
+use crate::rerender;
+
+// ------ RenderContext ------
+
+#[derive(Copy, Clone, Debug)]
+pub struct RenderContext {
+    pub index: u32,
+    pub node: LVar<Node>,
+    pub component_id: Option<TrackedCallId>,
+}
+
+impl RenderContext {
+    pub fn inc_index(&mut self) -> &mut Self {
+        self.index += 1;
+        self
+    } 
+
+    pub fn reset_index(&mut self) -> &mut Self {
+        self.index = 0;
+        self
+    } 
+
+    pub fn rerender(&self) {
+        rerender()
+    }
+}
